+++ history.md
@@ -6,6 +6,7 @@
 - Added a new package `meico.chroma` with classes `Chroma` and `Key`.
 - Added method `meico.msm.Msm.exportChroma()` that converts MSM to Chroma. It exports standard chroma features with 12 semitones in equal temperament and A = 440 Hz
 - The commandline app has a new option `-h` and `--chroma` to get chroma export. File `README.md` has been updated accordingly.
+- Redirected error messages to `System.err` instead of `System.out`. This makes the output in the commandline and log file more consistent.
 
 
 #### v0.3.0